requirements.txt
pydub
pygame
pillow